pyproject.toml
[project]
name = "rlm-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "dspy-ai>=3.1.3",
    "fastparquet>=2025.12.0",
    "fire>=0.7.1",
    "openinference-instrumentation-dspy>=0.1.33",
    "opentelemetry-exporter-otlp>=1.39.1",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "retry>=0.9.2",
    "rlms",
    "tqdm>=4.67.3",
]

[tool.uv.sources]
rlms = { git = "https://github.com/lambdaofgod/rlm" }

[tool.pytest.ini_options]
pythonpath = ["codeqa_comparison"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
